Control Rod Implementation
Reactor class modified to include a more complex heating solution, allowing for dynamic gameplay with the control rods and water cooling systems.

Control Rods modelled by chloebphillips@outlook.com, and successfully implemented in an Actor class, stabilizing the logical reaction and providing gameplay functionality with a valve.

SCRAM logically implemented as a failsafe, but with failure cases programmed in for later implementation, such as jammed rods, melting down, and an explosion.